fix(sessions): align archive retention CI contracts
