+++ docs/.generated/config-baseline.counts.json
@@ -1,5 +1,5 @@
 {
-  "core": 2418,
+  "core": 2419,
   "channel": 3710,
   "plugin": 4133
 }
