+++ docs/.generated/config-baseline.sha256
@@ -1,4 +1,4 @@
-f16c009f4f6c9975a27e397b19ca126ebc6d8f2dce111b0aa2845dc24ac5c32f  config-baseline.json
-5483d9fe277a5c5230ca43e412b7ac2d5a47c69be7ec8abcfc2eb02316a93204  config-baseline.core.json
+196c4f747071c633996bb25a9f3815e2e8632c511599136614bd4302ad17132d  config-baseline.json
+0b2457093a626f52d25396bbf74767bb905d36ec5b22195a9672b104687b0d20  config-baseline.core.json
 29d7abfcdefe9fbdf5c9e2eac31fefba0d6c8225e876ae03c9fedb62f4394e53  config-baseline.channel.json
 17bfcfce606c3d1830b2cebb78c970e2f8d9806af2e515b50ac608697a12ecb8  config-baseline.plugin.json
