+++ scripts/lib/session-accessor-debt-baseline.json
@@ -12,6 +12,7 @@
   },
   "sessionCompactManualTrim": {},
   "sessionLifecycleCleanup": {
+    "src/config/sessions/session-accessor.sqlite-maintenance.ts": 2,
     "src/config/sessions/session-accessor.sqlite-projection.ts": 2,
     "src/config/sessions/store-maintenance-operations.ts": 2,
     "src/infra/state-migrations.legacy-session-store.ts": 2
