+++ src/config/sessions/store.pruning.test.ts
@@ -21,7 +21,6 @@ import {
   pruneStaleEntries,
   pruneStaleModelRunEntries,
   resolveMaintenanceConfigFromInput,
-  normalizeResolvedMaintenanceConfigInput,
   resolveQuotaSuspensionEntryMaintenance,
   shouldPreserveMaintenanceEntry,
   shouldRunModelRunPrune,
@@ -30,22 +29,6 @@ import {
 import type { SessionEntry } from "./types.js";
 
 const DAY_MS = 24 * 60 * 60 * 1000;
-
-it("inherits deleted archive retention for older resolved callers but preserves explicit disablement", () => {
-  const base = {
-    mode: "enforce" as const,
-    pruneAfterMs: DAY_MS,
-    maxEntries: 100,
-    resetArchiveRetentionMs: 7 * DAY_MS,
-    maxDiskBytes: null,
-    highWaterBytes: null,
-  };
-  expect(normalizeResolvedMaintenanceConfigInput(base).deletedArchiveRetentionMs).toBe(7 * DAY_MS);
-  expect(
-    normalizeResolvedMaintenanceConfigInput({ ...base, deletedArchiveRetentionMs: null })
-      .deletedArchiveRetentionMs,
-  ).toBeNull();
-});
 
 const fixtureSuite = createFixtureSuite("openclaw-pruning-suite-");
 
@@ -364,44 +347,6 @@ describe("applyFileBackedSessionStoreMaintenance", () => {
     expect(onMaintenanceApplied).toHaveBeenCalledOnce();
     expect(warn).toHaveBeenCalledWith("session transcript archive retention cleanup failed", {
       error: String(cleanupError),
-    });
-  });
-
-  it("passes independent deleted and reset archive retention rules", async () => {
-    const now = Date.now();
-    const store = makeStore([
-      ["stale", { sessionId: "stale-session", updatedAt: now - 30 * DAY_MS }],
-      ["fresh", { sessionId: "fresh-session", updatedAt: now }],
-    ]);
-    const cleanupArchivedSessionTranscripts = vi.fn(async () => {});
-
-    await applyFileBackedSessionStoreMaintenance({
-      storePath: "/tmp/openclaw-sessions/sessions.json",
-      store,
-      maintenanceConfig: {
-        mode: "enforce",
-        pruneAfterMs: 7 * DAY_MS,
-        maxEntries: 500,
-        modelRunPruneAfterMs: DAY_MS,
-        resetArchiveRetentionMs: 14 * DAY_MS,
-        deletedArchiveRetentionMs: 3 * DAY_MS,
-        maxDiskBytes: null,
-        highWaterBytes: null,
-      },
-      log: { warn: () => {}, info: () => {} },
-      artifacts: {
-        archiveRemovedSessionTranscripts: async () => new Set(),
-        removeRemovedSessionTrajectoryArtifacts: async () => {},
-        cleanupArchivedSessionTranscripts,
-      },
-    });
-
-    expect(cleanupArchivedSessionTranscripts).toHaveBeenCalledWith({
-      directories: ["/tmp/openclaw-sessions"],
-      rules: [
-        { reason: "deleted", olderThanMs: 3 * DAY_MS },
-        { reason: "reset", olderThanMs: 14 * DAY_MS },
-      ],
     });
   });
 
@@ -1021,19 +966,8 @@ describe("resolveMaintenanceConfigFromInput", () => {
     });
 
     expect(maintenance.resetArchiveRetentionMs).toBe(7 * DAY_MS);
-    expect(maintenance.deletedArchiveRetentionMs).toBe(7 * DAY_MS);
     expect(maintenance.maxDiskBytes).toBeNull();
     expect(maintenance.highWaterBytes).toBeNull();
-  });
-
-  it("allows deleted archive retention to diverge from reset history", () => {
-    const maintenance = resolveMaintenanceConfigFromInput({
-      resetArchiveRetention: false,
-      deletedArchiveRetention: "30d",
-    });
-
-    expect(maintenance.resetArchiveRetentionMs).toBeNull();
-    expect(maintenance.deletedArchiveRetentionMs).toBe(30 * DAY_MS);
   });
 
   it("disables the disk budget when an explicit maxDiskBytes fails to parse", () => {
